3er Cambio
Se mejoró la lógica de los botones repetir y aleatorio, se configuró tema claro y oscuro, finalmente se organizó mejor el ViewModel.

+++ PlayerMusic/app/src/main/java/com/example/playermusic/ui/model/MusicModel.kt
@@ -3,7 +3,7 @@ package com.example.playermusic.ui.model
 data class MusicModel(
     //Datos de la música según el artista
     val musicPath: String = "",//Ruta del archivo de música
-    val musicDuration: Int = 0,//Duración total de la música
+    val musicDuration: Long = 0L,//Duración total de la música
     val musicName: String = "",
     val artistName: String = "",
     val albumName: String = "",
